+++ packages/components/molecules/f-rating/src/components/Rating.vue
@@ -2,32 +2,12 @@
     <div
         :class="$style['c-rating']"
         data-test-id="rating-component">
-        <div :class="$style['c-rating-starWrapper']">
-            <div :class="$style['c-rating-container']">
-                <star-icon
-                    v-for="star in maxStarRating"
-                    :key="star"
-                    :class="[
-                        $style['c-rating-star-empty'],
-                        $style[`c-rating-star--${starRatingSize}`]
-                    ]" />
-            </div>
-
-            <div
-                :class="[
-                    $style['c-rating-mask'],
-                    $style['c-rating-container']
-                ]"
-                :style="`--starRatingPercentage: ${getRatingStarPercentage}`">
-                <star-filled-icon
-                    v-for="star in maxStarRating"
-                    :key="star"
-                    :style="`--starRatingSize: c-rating-star--${starRatingSize}`"
-                    :class="[
-                        $style['c-rating-star-filled'],
-                        $style[`c-rating-star--${starRatingSize}`]
-                    ]" />
-            </div>
+        <div :class="$style['c-rating-stars']">
+            <component
+                :is="setRatingVariant"
+                :max-star-rating="maxStarRating"
+                :star-rating-size="starRatingSize"
+                :star-rating="starRating" />
 
             <span
                 data-test-id="c-rating-description"
@@ -35,23 +15,33 @@
                 {{ getRatingDescription }}
             </span>
         </div>
+
+        <span
+            v-if="ratingDisplayType"
+            data-test-id="c-rating-displayType"
+            :class="[
+                $style['c-rating-message'],
+                $style[`c-rating-message--${ratingDisplayType}`]]">
+            {{ getRatingDisplayFormat() }}
+        </span>
     </div>
 </template>
 
 <script>
-import {
-    StarFilledIcon,
-    StarIcon
-} from '@justeattakeaway/pie-icons-vue';
 import { VueGlobalisationMixin } from '@justeat/f-globalisation';
 import tenantConfigs from '../tenants';
-import { VALID_STAR_RATING_SIZES } from '../constants';
+import {
+    VALID_STAR_RATING_SIZES,
+    VALID_STAR_RATING_DISPLAY_TYPE
+} from '../constants';
+import RatingMultiStar from './RatingMultiStarVariant.vue';
+import RatingSingleStar from './RatingSingleStarVariant.vue';
 
 export default {
     name: 'VRating',
     components: {
-        StarFilledIcon,
-        StarIcon
+        RatingMultiStar,
+        RatingSingleStar
     },
 
     mixins: [VueGlobalisationMixin],
@@ -68,12 +58,26 @@ export default {
         },
         maxStarRating: {
             type: Number,
-            default: 5
+            default: 5,
+            validator: value => value > 0 && value % 1 === 0 // Positive integers
         },
         starRatingSize: {
             type: String,
             default: 'small',
-            validator: value => !!VALID_STAR_RATING_SIZES[value]
+            validator: value => VALID_STAR_RATING_SIZES.includes(value)
+        },
+        ratingDisplayType: {
+            type: String,
+            default: null,
+            validator: value => VALID_STAR_RATING_DISPLAY_TYPE.includes(value)
+        },
+        reviewCount: {
+            type: Number,
+            default: null
+        },
+        isSingleStarVariant: {
+            type: Boolean,
+            default: false
         }
     },
 
@@ -94,22 +98,42 @@ export default {
             return this.starRating < 2
                 ? this.$tc('ratings.starsDescription', 1, {
                     rating: this.starRating,
-                    total: this.maxStarRating
+                    maxStarRating: this.maxStarRating
                 })
 
                 : this.$tc('ratings.starsDescription', 2, {
                     rating: this.starRating,
-                    total: this.maxStarRating
+                    maxStarRating: this.maxStarRating
                 });
         },
 
         /**
-         * Calculate a percentage from the `starRating` value passed in by the consuming application.
+         * Return the correct imported star variant to use based on prop.
          *
          * @returns {string}
          */
-        getRatingStarPercentage () {
-            return `${(this.starRating / this.maxStarRating) * 100}%`;
+        setRatingVariant () {
+            return this.isSingleStarVariant
+                ? 'rating-single-star'
+                : 'rating-multi-star';
+        }
+    },
+
+    methods: {
+        /**
+         * Gets the correct rating display format from translations.
+         *
+         * @todo - If the component is using `short` as a `starRatingSize` we shouldn't display text
+         * alongside it for now. (TBC with design - ticket in backlog).
+         *
+         * @returns {string}
+         */
+        getRatingDisplayFormat () {
+            return this.$t(`ratings.ratingDisplayType.${this.ratingDisplayType}`, {
+                rating: this.starRating,
+                maxStarRating: this.maxStarRating,
+                reviewCount: this.reviewCount
+            });
         }
     }
 };
@@ -117,52 +141,13 @@ export default {
 
 <style lang="scss" module>
 @use '@justeat/fozzie/src/scss/fozzie' as f;
+@use '../../src/assets/scss/rating';
 
-.c-rating-starWrapper {
-    margin: 0;
-    padding: 0;
-    position: relative;
-    display: inline-block;
+.c-rating-message {
+    display: flex;
+    align-items: center;
+    @include f.font-size('body-s');
+    font-weight: f.$font-weight-bold;
+    margin-left: f.spacing(a);
 }
-    .c-rating-container {
-        display: flex;
-    }
-
-    .c-rating-star {
-        &--small {
-            width: 12px;
-        }
-
-        &--medium {
-            width: 16px;
-        }
-
-        &--large {
-            width: 28px;
-        }
-    }
-
-    .c-rating-star-filled {
-        & path {
-            fill: f.$color-support-brand-01;
-        }
-    }
-
-    .c-rating-star-empty {
-        & path {
-            fill: f.$color-mozzarella-50;
-        }
-    }
-
-    .c-rating-mask {
-        position: absolute;
-        top: 0;
-        left: 0;
-        overflow: hidden;
-        width: var(--starRatingPercentage);
-
-        svg {
-            flex-shrink: 0;
-        }
-    }
 </style>