+++ packages/components/pages/f-account-info/vue.config.js
@@ -17,6 +17,16 @@ module.exports = {
                 // eslint-disable-next-line quotes
                 additionalData: `@use "../assets/scss/common.scss";`
             });
+
+        config.externals({
+            '@justeat/f-alert': '@justeat/f-alert',
+            '@justeat/f-button': '@justeat/f-button',
+            '@justeat/f-card': '@justeat/f-card',
+            '@justeat/f-card-with-content': '@justeat/f-card-with-content',
+            '@justeat/f-error-message': '@justeat/f-error-message',
+            '@justeat/f-form-field': '@justeat/f-form-field',
+            '@justeat/f-link': '@justeat/f-link'
+        });
     },
     pluginOptions: {
         lintStyleOnBuild: true