+++ packages/components/pages/f-checkout/src/services/addressService/tenants/nz.js
@@ -6,6 +6,7 @@ function getEmptyAddress (postcode) {
     return {
         line1: '',
         line2: '',
+        line3: '',
         locality: '',
         postcode: postcode || ''
     };
@@ -15,6 +16,7 @@ function formatAddress (address) {
     return {
         line1: address.line1 || '',
         line2: address.line2 || '',
+        line3: address.line3 || '',
         locality: address.City || address.locality || '',
         postcode: address.postcode || address.PostalCode || address.ZipCode || address.postalCode || ''
     };