+++ packages/components/molecules/f-rating/CHANGELOG.md
@@ -3,6 +3,34 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.6.0
+------------------------------
+*October 26, 2022*
+
+### Added
+- rating.scss to shared styles between components.
+- Unit tests to cover changes.
+- Visual tests.
+
+### Changed
+- Split out from main rating file into Five & Single star components.
+-
+
+v0.5.0
+------------------------------
+*October 24, 2022*
+
+### Added
+- Test coverage for code changes.
+- Visual tests.
+- Prop validation.
+- Props `ratingDisplayType` & `reviewCount`.
+- Storybook contains ratings in different sizes & formats.
+
+### Changed
+- Class names to be more suitable.
+
+
 v0.4.0
 ------------------------------
 *October 19, 2022*